# [inductor][cpp][gemm] fix perf regression xcit_large_24_p8_224 (#134686)